collect all sparse tests in test_sparse.py
